Replace couch port build system
The old system was a hard to follow script. This
is now replaced by a more rebar friendly set of
files which tries to minimize dynamic config by
allowing a configure script to provide a
user.config file that overrides some of the port
environment.

Each driver now has a few separate environment
variables to allow proper build configuration
to be passed to the appropriate build steps
automatically. I've also ported the environment
configuration the old script produced
programatically but it should be tested and
adjusted on each platform.